fix: almost all types have JSDoc
